refactored MappingsFilterDialog
